requirements.txt
# FastAPI y servidor ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de datos
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Validación y configuración
pydantic==2.5.0
pydantic-settings==2.1.0

# Firebase Admin SDK
firebase-admin==6.4.0

# Google Cloud Secret Manager
google-cloud-secret-manager==2.18.0

# Utilidades
python-dotenv==1.0.0
